# docs/swagger.yaml
basePath: /api
definitions:
...
  attempt.SetupAttempt:
    properties:
      attach_to_self:
        type: boolean
      created_at:
        type: string
      customer:
        format: uuid
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      flow_directions:
        items:
          type: string
        type: array
      id:
        type: string
      livemode:
        type: boolean
      metadata:
        $ref: '#/definitions/core.JSONB'
      object:
        type: string
      payment_method:
        format: uuid
        type: string
      payment_method_details:
        format: uuid
        type: string
      setup_error:
        format: uuid
        type: string
      setup_intent:
        format: uuid
        type: string
      status:
        $ref: '#/definitions/attempt.Status'
      updated_at:
        type: string
      usage:
        $ref: '#/definitions/attempt.Usage'
    type: object
  attempt.Status:
    enum:
    - requires_confirmation
    - requires_action
    - processing
    - succeeded
    - failed
    - abandoned
    type: string
    x-enum-varnames:
    - StatusRequiresConfirmation
    - StatusRequiresAction
    - StatusProcessing
    - StatusSucceeded
    - StatusFailed
    - StatusAbandoned
  attempt.Usage:
    enum:
    - off_session
    - on_session
    type: string
    x-enum-varnames:
    - UsageOffSession
    - UsageOnSession
...
  core.JSONB:
    additionalProperties: true
    type: object
...
  database.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  mandate.Mandate:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      id:
        type: string
      livemode:
        type: boolean
      metadata:
        $ref: '#/definitions/core.JSONB'
      object:
        type: string
      payment_method:
        format: uuid
        type: string
      payment_method_details:
        format: uuid
        type: string
      single_use:
        format: uuid
        type: string
      status:
        $ref: '#/definitions/mandate.Status'
      type:
        $ref: '#/definitions/mandate.Type'
      updated_at:
        type: string
    type: object
  mandate.Status:
    enum:
    - active
    - inactive
    - pending
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusPending
  mandate.Type:
    enum:
    - multi_use
    - single_use
    type: string
    x-enum-varnames:
    - TypeMultiUse
    - TypeSingleUse
...
  payout.Method:
    enum:
    - standard
    - instant
    type: string
    x-enum-varnames:
    - MethodStandard
    - MethodInstant
  payout.Payout:
    properties:
      amount:
        type: integer
      arrival_date:
        type: integer
      automatic:
        type: boolean
      balance_transaction:
        format: uuid
        type: string
      created_at:
        type: string
      currency:
        type: string
      deleted_at:
        $ref: '#/definitions/database.DeletedAt'
      description:
        type: string
      destination:
        type: string
      failure_balance_transaction:
        format: uuid
        type: string
      failure_code:
        type: string
      failure_message:
        type: string
      id:
        type: string
      livemode:
        type: boolean
      metadata:
        $ref: '#/definitions/core.JSONB'
      method:
        $ref: '#/definitions/payout.Method'
      object:
        type: string
      original_payout:
        format: uuid
        type: string
      reversed_by:
        format: uuid
        type: string
      source_type:
        $ref: '#/definitions/payout.SourceType'
      statement_descriptor:
        type: string
      status:
        $ref: '#/definitions/payout.Status'
      type:
        $ref: '#/definitions/payout.Type'
      updated_at:
        type: string
    type: object
  payout.SourceType:
    enum:
    - card
    - fpx
    - bank_account
    type: string
    x-enum-varnames:
    - SourceTypeCard
    - SourceTypeFpx
    - SourceTypeBankAccount
  payout.Status:
    enum:
    - paid
    - pending
    - in_transit
    - canceled
    - failed
    type: string
    x-enum-varnames:
    - StatusPaid
    - StatusPending
    - StatusInTransit
    - StatusCanceled
    - StatusFailed
  payout.Type:
    enum:
    - bank_account
    - card
    type: string
    x-enum-varnames:
    - TypeBankAccount
    - TypeCard
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
...
  /v1/mandates/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Mandate by given ID.
      parameters:
      - description: Mandate ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete Mandate by given ID
      tags:
      - Mandate
    get:
      consumes:
      - application/json
      description: Get Mandate by given ID or 404 error.
      parameters:
      - description: Mandate ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mandate.Mandate'
      summary: get Mandate by given ID or 404 error.
      tags:
      - Mandate
    post:
      consumes:
      - application/json
      description: Update Mandate.
      parameters:
      - description: Mandate ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mandate.Mandate'
      summary: update Mandate
      tags:
      - Mandate
...
  /v1/payouts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Payout by given ID.
      parameters:
      - description: Payout ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete Payout by given ID
      tags:
      - Payout
    get:
      consumes:
      - application/json
      description: Get Payout by given ID or 404 error.
      parameters:
      - description: Payout ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Payout'
      summary: get Payout by given ID or 404 error.
      tags:
      - Payout
    post:
      consumes:
      - application/json
      description: Update Payout.
      parameters:
      - description: Payout ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Payout'
      summary: update Payout
      tags:
      - Payout
  /v1/payouts/{id}/cancele:
    post:
      consumes:
      - application/json
      description: Cancel Payout.
      parameters:
      - description: Payout ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Payout'
      summary: cancel Payout
      tags:
      - Payout
  /v1/payouts/{id}/reverse:
    post:
      consumes:
      - application/json
      description: Reverse Payout.
      parameters:
      - description: Payout ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Payout'
      summary: reverse Payout
      tags:
      - Payout
...
  /v1/setup_attempts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete SetupAttempt by given ID.
      parameters:
      - description: SetupAttempt ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "204":
          description: ok
          schema:
            type: string
      summary: delete SetupAttempt by given ID
      tags:
      - SetupAttempt
    get:
      consumes:
      - application/json
      description: Get SetupAttempt by given ID or 404 error.
      parameters:
      - description: SetupAttempt ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/attempt.SetupAttempt'
      summary: get SetupAttempt by given ID or 404 error.
      tags:
      - SetupAttempt
    post:
      consumes:
      - application/json
      description: Update SetupAttempt.
      parameters:
      - description: SetupAttempt ID
        in: body
        name: id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/attempt.SetupAttempt'
      summary: update SetupAttempt
      tags:
      - SetupAttempt
...
schemes:
- http
swagger: "2.0"